+++ examples/document-search/qdrant.py
@@ -94,7 +94,7 @@ async def main() -> None:
     query = "I'm boiling my water and I need a joke"
     vector_store_kwargs = {
         "k": 2,
-        "max_distance": 0.6,
+        "score_threshold": 0.6,
     }
     results = await document_search.search(
         query,