+++ packages/ragbits-core/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 - Add HybridSearchVectorStore which can aggregate results from multiple VectorStores (#412)
+- Make the score in VectorStoreResult consistent (always bigger is better)
 
 ## 0.10.2 (2025-03-21)